Contains space optimized DP code
If we observe carefully, we can see that the dp solution with states (i,j)
will depend on state (i-1, j) or (i-1, j-wt[i-1]). In either case the solution
for state (i,j) will lie in the i-1th row of the memoization table. So at
every iteration of the index, we can copy the values of current row and
use only this row for building the solution in next iteration and no
other row will be used. Hence, at any iteration we will be using only a
single row to build the solution for current row. Hence, we can reduce
the space complexity to just O(W).