CodSule July'19: Connecting 2-D coordinates in a grid - Anti-clockwise - Priority Queue
You are given a grid with n rows and m columns. Each of the numbers from 1 to x is present twice in the grid, at the coordinates (ai,bi) and (ci,di).

You have to calculate if it is possible to draw curves that join the two pairs of coordinates for each of the numbers, in such a way that none of the curves overlap, or leave the boundaries of the grid.

We call a number 'critical', if both occurrences of this number is on the border of the grid. Suppose that there are two 
critical numbers a and b, and they appear in the border in the order a, b, a, b. In this case, it is clear that we can’t 
connect both numbers without overlapping, thus the answer is 'No'. Otherwise, we can connect all numbers in the following 
order:

First, we connect non-critical numbers one by one.
Then, we connect critical numbers one by one. This way, we never get stuck.

How can we check if there exists such a pair (a, b) of critical numbers? Start from any point on the border and check all 
critical points on the border one by one in clockwise order (ignore non-critical numbers on the border). Initially we have 
an empty stack, and whenever we find an critical number x on the border, do the following:

If the top element of the stack is also x, pop it.
Otherwise, push x into the stack.

If the stack becomes empty after we check all points, the answer is 'Yes'. Otherwise the answer is 'No'. 
This algorithm works in O(NlogN).