Hackerrarnk: Similar to Sprinkler Interview Question
Given a string consisting of a, b and c's, we can perform the following operation: Take any two adjacent distinct characters and replace it with the third character. For example, if 'a' and 'c' are adjacent, they can be replaced with 'b'. What is the smallest string which can result by applying this operation repeatedly?